+++ demo/lib/cases/case_builder.dart
@@ -3,7 +3,6 @@ import 'package:demo/case.dart';
 import 'package:demo/helper.dart';
 import 'package:flexiblebox/flexiblebox_flutter.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 
 final GlobalKey key3 = GlobalKey();